bug: 26812
p4 integrate -b FRANKLIN_to_main @106802,106802
derby seems to be prone to corruption if not explicitly shutdown.  to minimize chance of corruption, we do explicity shutdown from Zimbra.shutdown().

http://bugzilla.zimbra.com/show_bug.cgi?id=26812

+++ ZimbraOffline/src/java/com/zimbra/cs/account/offline/DirectorySync.java
@@ -44,6 +44,7 @@ import com.zimbra.cs.offline.OfflineSyncManager;
 import com.zimbra.cs.service.account.ModifyPrefs;
 import com.zimbra.cs.servlet.ZimbraServlet;
 import com.zimbra.cs.util.Zimbra;
+import com.zimbra.cs.util.ZimbraApplication;
 import com.zimbra.cs.zclient.ZFilterRules;
 import com.zimbra.cs.zclient.ZGetInfoResult;
 import com.zimbra.cs.zclient.ZIdentity;
@@ -62,6 +63,8 @@ public class DirectorySync {
 	private DirectorySync() {
 		timer.schedule(new TimerTask() {
 				public void run() {
+					if (ZimbraApplication.getInstance().isShutdown())
+						return;
 					try {
 						syncAllAccounts(false);
 					} catch (Throwable e) { //don't let exceptions kill the timer
